+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - `Searcher.sorted_search` to obtain a list of sorted `SearchResult`s.
 - `grascii.dictionary.common` module to contain `DictionaryException`s and utility functions.
 - `Dictionary` class to work with grascii dictionaries.
+- `config.get_default_config` to get the text of the default configuration file.
 
 ### Changed